// bachelor-front-react-version-main/src/generated-sources/openapi/api.ts
export interface Business {
  /**
   *
   * @type {string}
   * @memberof Business
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof Business
   */
  ownerUserId?: string;
  /**
   *
   * @type {BusinessState}
   * @memberof Business
   */
  businessState?: BusinessState;
  /**
   *
   * @type {BusinessDescription}
   * @memberof Business
   */
  businessDescription: BusinessDescription;
  /**
   *
   * @type {BusinessPaymentDetails}
   * @memberof Business
   */
  businessPaymentDetails: BusinessPaymentDetails;
  /**
   *
   * @type {BusinessLegalDocuments}
   * @memberof Business
   */
  businessLegalDocuments: BusinessLegalDocuments;
  /**
   *
   * @type {BusinessContacts}
   * @memberof Business
   */
  businessContacts?: BusinessContacts;
  /**
   *
   * @type {BusinessLogo}
   * @memberof Business
   */
  businessLogo?: BusinessLogo;
  /**
   *
   * @type {string}
   * @memberof Business
   */
  createdDate?: string;
  /**
   *
   * @type {string}
   * @memberof Business
   */
  lastModifiedDate?: string;
}
/**
 *
 * @export
 * @interface BusinessApplication
 */
export interface BusinessApplication {
  /**
   *
   * @type {string}
   * @memberof BusinessApplication
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof BusinessApplication
   */
  createdBusinessId?: string;
  /**
   *
   * @type {string}
   * @memberof BusinessApplication
   */
  creatorUserId?: string;
  /**
   *
   * @type {string}
   * @memberof BusinessApplication
   */
  reviewId?: string;
  /**
   *
   * @type {BusinessApplicationState}
   * @memberof BusinessApplication
   */
  businessApplicationState?: BusinessApplicationState;
  /**
   *
   * @type {BusinessDescription}
   * @memberof BusinessApplication
   */
  businessDescription: BusinessDescription;
  /**
   *
   * @type {BusinessPaymentDetails}
   * @memberof BusinessApplication
   */
  businessPaymentDetails: BusinessPaymentDetails;
  /**
   *
   * @type {BusinessLegalDocuments}
   * @memberof BusinessApplication
   */
  businessLegalDocuments: BusinessLegalDocuments;
  /**
   *
   * @type {BusinessContacts}
   * @memberof BusinessApplication
   */
  businessContacts?: BusinessContacts;
  /**
   *
   * @type {BusinessLogo}
   * @memberof BusinessApplication
   */
  businessLogo?: BusinessLogo;
}
// ...
export interface BusinessApplicationReview {
  /**
   *
   * @type {string}
   * @memberof BusinessApplicationReview
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof BusinessApplicationReview
   */
  BusinessApplicationId?: string;
  /**
   *
   * @type {string}
   * @memberof BusinessApplicationReview
   */
  ReviewAuthorId?: string;
  /**
   *
   * @type {BusinessReviewState}
   * @memberof BusinessApplicationReview
   */
  businessReviewState?: BusinessReviewState;
  /**
   *
   * @type {string}
   * @memberof BusinessApplicationReview
   */
  businessReviewDescription?: string;
}
/**
 *
 * @export
 * @enum {string}
 */

export const BusinessApplicationState = {
  InReview: "InReview",
  Approved: "Approved",
  Denied: "Denied",
} as const;

export type BusinessApplicationState =
  (typeof BusinessApplicationState)[keyof typeof BusinessApplicationState];

/**
 *
 * @export
 * @interface BusinessContacts
 */
export interface BusinessContacts {
  /**
   *
   * @type {number}
   * @memberof BusinessContacts
   */
  phoneNumber?: number;
  /**
   *
   * @type {string}
   * @memberof BusinessContacts
   */
  email?: string;
  /**
   *
   * @type {string}
   * @memberof BusinessContacts
   */
  address?: string;
  /**
   *
   * @type {string}
   * @memberof BusinessContacts
   */
  website?: string;
  /**
   *
   * @type {string}
   * @memberof BusinessContacts
   */
  facebook?: string;
  /**
   *
   * @type {string}
   * @memberof BusinessContacts
   */
  instagram?: string;
}
/**
 *
 * @export
 * @interface BusinessDescription
 */
export interface BusinessDescription {
  /**
   *
   * @type {string}
   * @memberof BusinessDescription
   */
  legalName: string;
  /**
   *
   * @type {string}
   * @memberof BusinessDescription
   */
  businessEntityType: BusinessDescriptionBusinessEntityTypeEnum;
  /**
   *
   * @type {number}
   * @memberof BusinessDescription
   */
  uniqueIdentificationCode: number;
  /**
   *
   * @type {number}
   * @memberof BusinessDescription
   */
  tvaNumber: number;
  /**
   *
   * @type {string}
   * @memberof BusinessDescription
   */
  companyDescription?: string;
}
// ...
export interface BusinessLegalDocuments {
  /**
   * File key from AWS S3 bucket
   * @type {string}
   * @memberof BusinessLegalDocuments
   */
  registrationCertificateKey: string;
  /**
   * AWS S3 Object Url to bank statement file
   * @type {string}
   * @memberof BusinessLegalDocuments
   */
  registrationCertificateFileUrl: string;
  /**
   * File key from AWS S3 bucket
   * @type {string}
   * @memberof BusinessLegalDocuments
   */
  bankStatementFileKey: string;
  /**
   * AWS S3 Object Url to bank statement file
   * @type {string}
   * @memberof BusinessLegalDocuments
   */
  bankStatementFileUrl: string;
}
/**
 *
 * @export
 * @interface BusinessLogo
 */
export interface BusinessLogo {
  /**
   * File key from AWS S3 bucket
   * @type {string}
   * @memberof BusinessLogo
   */
  logoKey: string;
  /**
   * AWS S3 Object Url to bank statement file
   * @type {string}
   * @memberof BusinessLogo
   */
  logoFileUrl: string;
}
/**
 *
 * @export
 * @interface BusinessPaymentDetails
 */
export interface BusinessPaymentDetails {
  /**
   *
   * @type {string}
   * @memberof BusinessPaymentDetails
   */
  iban: string;
  /**
   *
   * @type {string}
   * @memberof BusinessPaymentDetails
   */
  bank: string;
  /**
   *
   * @type {string}
   * @memberof BusinessPaymentDetails
   */
  swiftCode: string;
}
/**
 *
 * @export
 * @enum {string}
 */

export const BusinessReviewState = {
  Approved: "Approved",
  Denied: "Denied",
} as const;

export type BusinessReviewState =
  (typeof BusinessReviewState)[keyof typeof BusinessReviewState];

/**
 *
 * @export
 * @enum {string}
 */

export const BusinessState = {
  Active: "Active",
  Blocked: "Blocked",
} as const;

export type BusinessState = (typeof BusinessState)[keyof typeof BusinessState];
// ...
export interface CreateBusinessApplicationRequest {
  /**
   *
   * @type {BusinessDescription}
   * @memberof CreateBusinessApplicationRequest
   */
  businessDescription: BusinessDescription;
  /**
   *
   * @type {BusinessPaymentDetails}
   * @memberof CreateBusinessApplicationRequest
   */
  businessPaymentDetails: BusinessPaymentDetails;
  /**
   *
   * @type {BusinessLegalDocuments}
   * @memberof CreateBusinessApplicationRequest
   */
  businessLegalDocuments: BusinessLegalDocuments;
  /**
   *
   * @type {BusinessContacts}
   * @memberof CreateBusinessApplicationRequest
   */
  businessContacts?: BusinessContacts;
  /**
   *
   * @type {BusinessLogo}
   * @memberof CreateBusinessApplicationRequest
   */
  businessLogo?: BusinessLogo;
}
// ...
export interface CreateCategoryApplicationRequest {
  /**
   *
   * @type {string}
   * @memberof CreateCategoryApplicationRequest
   */
  category?: string;
}
// ...
export const OrderStatus = {
  AwaitingShipment: "AWAITING_SHIPMENT",
  ReadyForShipment: "READY_FOR_SHIPMENT",
  Shipped: "SHIPPED",
  AwaitingPickup: "AWAITING_PICKUP",
  Completed: "COMPLETED",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

/**
 *
 * @export
 * @interface OrderToFulfill
 */
export interface OrderToFulfill {
  /**
   *
   * @type {string}
   * @memberof OrderToFulfill
   */
  id?: string;
  /**
   *
   * @type {Array<ProductInOrder>}
   * @memberof OrderToFulfill
   */
  productList?: Array<ProductInOrder>;
  /**
   *
   * @type {string}
   * @memberof OrderToFulfill
   */
  fulfilledBy?: string;
  /**
   *
   * @type {string}
   * @memberof OrderToFulfill
   */
  orderedBy?: string;
  /**
   *
   * @type {ShippingDetails}
   * @memberof OrderToFulfill
   */
  shippingDetails?: ShippingDetails;
  /**
   *
   * @type {string}
   * @memberof OrderToFulfill
   */
  createdDate?: string;
  /**
   *
   * @type {OrderStatus}
   * @memberof OrderToFulfill
   */
  orderStatus?: OrderStatus;
}
// ...
export interface PatchOrderToFulfillProductStatusRequest {
  /**
   *
   * @type {string}
   * @memberof PatchOrderToFulfillProductStatusRequest
   */
  orderId: string;
  /**
   *
   * @type {string}
   * @memberof PatchOrderToFulfillProductStatusRequest
   */
  productId: string;
  /**
   *
   * @type {OrderStatus}
   * @memberof PatchOrderToFulfillProductStatusRequest
   */
  status: OrderStatus;
}
/**
 *
 * @export
 * @interface PatchOrderToFulfillStatusRequest
 */
export interface PatchOrderToFulfillStatusRequest {
  /**
   *
   * @type {string}
   * @memberof PatchOrderToFulfillStatusRequest
   */
  orderId: string;
  /**
   *
   * @type {OrderStatus}
   * @memberof PatchOrderToFulfillStatusRequest
   */
  status: OrderStatus;
}
/**
 *
 * @export
 * @interface PatchProductRequest
 */
export interface PatchProductRequest {
  /**
   *
   * @type {string}
   * @memberof PatchProductRequest
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof PatchProductRequest
   */
  name?: string;
  /**
   *
   * @type {number}
   * @memberof PatchProductRequest
   */
  price?: number;
  /**
   *
   * @type {string}
   * @memberof PatchProductRequest
   */
  categoryId?: string;
  /**
   *
   * @type {string}
   * @memberof PatchProductRequest
   */
  description?: string;
  /**
   *
   * @type {number}
   * @memberof PatchProductRequest
   */
  stockAmount?: number;
  /**
   *
   * @type {ProductImage}
   * @memberof PatchProductRequest
   */
  mainImage?: ProductImage;
  /**
   *
   * @type {Array<ProductImage>}
   * @memberof PatchProductRequest
   */
  additionalImages?: Array<ProductImage>;
}
/**
 *
 * @export
 * @interface PrivateBusinessAnalytics
 */
export interface PrivateBusinessAnalytics {
  /**
   *
   * @type {string}
   * @memberof PrivateBusinessAnalytics
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof PrivateBusinessAnalytics
   */
  belongsToBusinessId?: string;
  /**
   *
   * @type {number}
   * @memberof PrivateBusinessAnalytics
   */
  totalNumberOfCompletedOrders?: number;
  /**
   *
   * @type {number}
   * @memberof PrivateBusinessAnalytics
   */
  totalNumberOfOrdersToBeCompleted?: number;
  /**
   *
   * @type {number}
   * @memberof PrivateBusinessAnalytics
   */
  totalIncome?: number;
  /**
   *
   * @type {number}
   * @memberof PrivateBusinessAnalytics
   */
  totalSoldAmountProducts?: number;
}
/**
 *
 * @export
 * @interface Product
 */
export interface Product {
  /**
   *
   * @type {string}
   * @memberof Product
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  belongsToBusinessId?: string;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  categoryId?: string;
  /**
   *
   * @type {number}
   * @memberof Product
   */
  price?: number;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  description?: string;
  /**
   *
   * @type {ProductRating}
   * @memberof Product
   */
  rating?: ProductRating;
  /**
   *
   * @type {string}
   * @memberof Product
   */
  listedDate?: string;
  /**
   *
   * @type {number}
   * @memberof Product
   */
  stockAmount?: number;
  /**
   *
   * @type {number}
   * @memberof Product
   */
  totalSold?: number;
  /**
   *
   * @type {ProductImage}
   * @memberof Product
   */
  mainImage?: ProductImage;
  /**
   *
   * @type {Array<ProductImage>}
   * @memberof Product
   */
  additionalImages?: Array<ProductImage>;
}
// ...
export interface ProductRequest {
  /**
   *
   * @type {string}
   * @memberof ProductRequest
   */
  name: string;
  /**
   *
   * @type {number}
   * @memberof ProductRequest
   */
  price: number;
  /**
   *
   * @type {string}
   * @memberof ProductRequest
   */
  categoryId?: string;
  /**
   *
   * @type {string}
   * @memberof ProductRequest
   */
  description: string;
  /**
   *
   * @type {number}
   * @memberof ProductRequest
   */
  stockAmount: number;
  /**
   *
   * @type {ProductImage}
   * @memberof ProductRequest
   */
  mainImage: ProductImage;
  /**
   *
   * @type {Array<ProductImage>}
   * @memberof ProductRequest
   */
  additionalImages?: Array<ProductImage>;
}
// ...
export interface UpdateBusinessApplicationRequest {
  /**
   *
   * @type {string}
   * @memberof UpdateBusinessApplicationRequest
   */
  id: string;
  /**
   *
   * @type {BusinessDescription}
   * @memberof UpdateBusinessApplicationRequest
   */
  businessDescription: BusinessDescription;
  /**
   *
   * @type {BusinessPaymentDetails}
   * @memberof UpdateBusinessApplicationRequest
   */
  businessPaymentDetails: BusinessPaymentDetails;
  /**
   *
   * @type {BusinessLegalDocuments}
   * @memberof UpdateBusinessApplicationRequest
   */
  businessLegalDocuments: BusinessLegalDocuments;
  /**
   *
   * @type {BusinessContacts}
   * @memberof UpdateBusinessApplicationRequest
   */
  businessContacts?: BusinessContacts;
  /**
   *
   * @type {BusinessLogo}
   * @memberof UpdateBusinessApplicationRequest
   */
  businessLogo?: BusinessLogo;
}